Add x11.pc pseudo-package for macOS
The x11.pc file is missing on macOS (at least on Brew), and it was
getting annoying to add ad-hoc rules in every makefile that required
it. The hack here is to add one directly from make-it-quick.

Putting the make-it-quick directory last should ensure this solution
still works correclty if you have an actual x11.pc on your macOS system.