Fix regression in 'make test'
The regression was introduced in 96cad841f883efe6ad38917e5482f1b633daf86a
on Cygwin and macOS. The MIQ_DLLNAME variable used $(@F) instead of $@.

The Jenkins setup failed to see the problem because it rebuilt
on top of existing directories that were pruned with existing libraries.

Added a second build after `git clean -dfx` in Jenkins to catch that.