Place -I options from configuration behind project's
Discovered that while compiling on a machine where some package had
installed a header file in `/usr/local/include` with the same name as
a header in the project.

With the old rules, the command line looked like:

    -I/usr/local/include ... -I/project/include

where the first option came from some pkg-config output.

This caused compilation-errors, since the `/usr/local/include` header
was preferred over the project's.